requirements.txt
# Core dependencies for Cluster Bus Fuzzer
valkey[libvalkey]>=6.0.0  # For Valkey client connectivity with compiled response parser
PyYAML>=6.0  # For DSL configuration parsing

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Code quality
ruff>=0.1.0
black>=23.0.0
isort>=5.12.0